refactor(engine): LIFO bölge mantığı LifoPlacement.cs'e ayrıldı (ALG-04)
ComputeGroupZones ve bölge cezası hesabı ayrı dosyaya çıkarıldı; ComputeScore
artık LifoPlacement.ZonePenalty'yi çağırıyor. Katsayı (2000) ve Math.Max
sırası aynen korundu.

LIFO yön semantiği üç yerde örtük yaşıyordu (grup sıralaması, bölge sıralaması,
istif kuralı). CompareUnloadingOrder yardımcısı eklendi ve istif kuralı buna
bağlandı; diğer iki nokta çapraz referans yorumuyla işaretlendi (grup sıralaması
5. adımda ItemOrdering'e taşınacak).

Not: kararın bu adımda öngördüğü fantom bölge cezası düzeltmesi ve LIFO yön
tersliği dev dalında zaten mevcut; karar metni eski bir main kopyasına göre
yazılmıştı. Bu commit yalnızca yapısal ayrıştırmadır.

Davranış sıfır değişim: snapshot'ların hiçbiri değişmedi. Passed 23 + 20.
